simulator/requirements.txt
numpy==2.0.1
paho-mqtt==2.1.0
